% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.mr.plotly.R
\name{all.mr.plotly}
\alias{all.mr.plotly}
\title{Create RDS object containing all plots}
\usage{
\method{all}{mr.plotly}(
  res,
  res_single,
  res_loo,
  dat,
  out = NA,
  include_data = F,
  failed = NA,
  versions = NA
)
}
\arguments{
\item{res}{Output of MR analysis using TwoSampleMR \code{\link[TwoSampleMR]{mr}} function}

\item{res_single}{Output data.frame from \code{\link[TwoSampleMR]{mr_singlesnp}} or filename containing the result in csv format}

\item{res_loo}{Output of leave-one-out analyais from \code{\link[TwoSampleMR]{mr_leaveoneout}} function}

\item{dat}{Harmonized data. Output of the \code{\link[TwoSampleMR]{harmonise_data}} function from TwoSampleMR package}

\item{out}{Output file name. If provided the plots will be saved as .rds otherwise a list of all plots will be returned.}

\item{include_data}{Logical. Default is FALSE}

\item{failed}{A csv or tsv file containing information about failed runs. THe information will be displayed in the dashboard as a table.}

\item{version}{Software versions used to run the analysis to be displayed in the dashboard. Use \code{sessionInfo()} to generate the version information.}
}
\description{
Create RDS object containing all plots
}
\author{
Subrata Paul

bmi_exp_dat <- extract_instruments(outcomes = c('ieu-a-2', 'ieu-a-9', 'ieu-a-12'))
chd_out_dat <- extract_outcome_data(snps = bmi_exp_dat$SNP, outcomes = 'ieu-a-7')
dat <- harmonise_data(bmi_exp_dat, chd_out_dat)
res_single <- mr_singlesnp(dat)
all.mr.plotly(res_single)
}
