% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.mr.plotly.R
\name{all.rm.plotly}
\alias{all.rm.plotly}
\title{Create RDS object containing all plots}
\usage{
\method{all}{rm.plotly}(res, res_single, dat, out = NA, include_data = F)
}
\arguments{
\item{res_single}{Output data.frame from `mr_singlesnp` or filename containing the result in csv format}

\item{`res`}{- Output of MR analysis using TwoSampleMR `mr` function}

\item{`dat`}{- Harmonized data. Output of the `harmonise_data` function from TwoSampleMR package}

\item{`out`}{- Output file name. If provided the plots will be saved as .rds otherwise a list of all plots will be returned.}

\item{`include_data`}{- Logical. Default is FALSE}
}
\description{
Create RDS object containing all plots
}
\author{
Subrata Paul

bmi_exp_dat <- extract_instruments(outcomes = c('ieu-a-2', 'ieu-a-9', 'ieu-a-12'))
chd_out_dat <- extract_outcome_data(snps = bmi_exp_dat$SNP, outcomes = 'ieu-a-7')
dat <- harmonise_data(bmi_exp_dat, chd_out_dat)
res_single <- mr_singlesnp(dat)
all.mr.plotly(res_single)
}
