% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_plot.R
\name{blank_plot}
\alias{blank_plot}
\title{Creates a blank plot with provided message}
\usage{
blank_plot(message)
}
\description{
Creates a blank plot with provided message
}
\author{
Subrata Paul

This function is used to create a blank plot with a text message for the dashboard for the cases where data were not available or the TwoSampleMR model did not run for lack of SNPs
}
