% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.plotly.single.R
\name{forest.plotly.single}
\alias{forest.plotly.single}
\title{Interactive forest plot for single SNP two sample MR}
\usage{
forest.plotly.single(res_single, nticks = 40)
}
\arguments{
\item{res_single}{Output data.frame from `mr_singlesnp` or filename containing the result in csv format}
}
\description{
Interactive forest plot for single SNP two sample MR
}
\examples{
bmi_exp_dat <- extract_instruments(outcomes = c('ieu-a-2', 'ieu-a-9', 'ieu-a-12'))
chd_out_dat <- extract_outcome_data(snps = bmi_exp_dat$SNP, outcomes = 'ieu-a-7')
dat <- harmonise_data(bmi_exp_dat, chd_out_dat)
res_single <- mr_singlesnp(dat)
forest.plotly.single(res_single)

This function gives plotly object based on the results of `mr_singlesnp`. The forest plot compares the MR estimates using the different MR methods against the single SNP tests.
}
\author{
Subrata Paul
}
