% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel.plotly.R
\name{funnel.plotly}
\alias{funnel.plotly}
\title{Funnel plot for TwoSampleMR}
\usage{
funnel.plotly(res_single)
}
\arguments{
\item{`res_single`}{}
}
\description{
Funnel plot for TwoSampleMR
}
\examples{

bmi_exp_dat <- extract_instruments(outcomes = c('ieu-a-2', 'ieu-a-9', 'ieu-a-12'))
chd_out_dat <- extract_outcome_data(snps = bmi_exp_dat$SNP, outcomes = 'ieu-a-7')
dat <- harmonise_data(bmi_exp_dat, chd_out_dat)
res_single <- mr_singlesnp(dat)
funnel.plotly(res_single)

}
\author{
Subrata Paul
}
