% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.dashboard.R
\name{make.dashboard}
\alias{make.dashboard}
\title{Write a Rmd file for flex-dashboard}
\usage{
make.dashboard(file = "dashboard")
}
\arguments{
\item{file}{Filename to shave the dashboard .Rmd File}
}
\description{
To run the dashboard open the produced .Rmd file in RStudio and hit 'Run Documents'
}
\author{
Subrata Paul
}
