% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.loo.plotly.R
\name{res.loo.plotly}
\alias{res.loo.plotly}
\title{Plot TwoSampleMR leave-one-out crossvalidation}
\usage{
res.loo.plotly(res_loo)
}
\arguments{
\item{`res_loo`, }{output from `mr_leaveoneout`}
}
\description{
Plot TwoSampleMR leave-one-out crossvalidation
}
\examples{
bmi_exp_dat <- extract_instruments(outcomes = c('ieu-a-2', 'ieu-a-9', 'ieu-a-12'))
chd_out_dat <- extract_outcome_data(snps = bmi_exp_dat$SNP, outcomes = 'ieu-a-7')
dat <- harmonise_data(bmi_exp_dat, chd_out_dat)
res_loo <- mr_leaveoneout(dat)
res.loo.plotly(res_loo)

}
\author{
Subrata Paul
}
