% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.plotly.R
\name{scatter.plotly}
\alias{scatter.plotly}
\title{Scatter plot for TwoSampleMR analysis}
\usage{
scatter.plotly(res, dat)
}
\arguments{
\item{`res`}{- Output of MR analysis using TwoSampleMR `mr` function}

\item{`dat`}{- Harmonized data. Output of the `harmonise_data` function from TwoSampleMR package}
}
\description{
Scatter plot for TwoSampleMR analysis
}
\examples{

bmi_exp_dat <- extract_instruments(outcomes = c('ieu-a-2', 'ieu-a-9', 'ieu-a-12'))
chd_out_dat <- extract_outcome_data(snps = bmi_exp_dat$SNP, outcomes = 'ieu-a-7')
dat <- harmonise_data(bmi_exp_dat, chd_out_dat)
res = mr(dat)
scatter.plotly(res, dat)

}
\author{
Subrata Paul
}
