% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel.plotly.R
\name{funnel.plotly}
\alias{funnel.plotly}
\title{Funnel plot for TwoSampleMR}
\usage{
funnel.plotly(res_single)
}
\arguments{
\item{`res_single`}{}
}
\description{
Funnel plot for TwoSampleMR
}
\author{
Subrata Paul
}
