% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.loo.plotly.R
\name{res.loo.plotly}
\alias{res.loo.plotly}
\title{Plot TwoSampleMR leave-one-out crossvalidation}
\usage{
res.loo.plotly(res_loo)
}
\arguments{
\item{`res_loo`, }{output from `mr_leaveoneout`}
}
\description{
Plot TwoSampleMR leave-one-out crossvalidation
}
\author{
Subrata Paul
}
